Added random patrol circular area command

+++ GameV1/CombatHandler.cs
@@ -8,7 +8,7 @@ namespace GameV1
         {
             int damage = (int)(attacker.Stats.Strength + attackWeapon.Damage);
             //int damage = (int)(attacker.Stats.Strength + attacker.StrongestWeapon.Damage);
-            double damageModifier = 100.0f / (100.0f + ((double)defender.Chest.Item.DamageReduction * (1.0f - ((double)attackWeapon.ArmorPenetrationPercent / 100.0f)) - (double)attackWeapon.ArmorPenetrationFlat));
+            float damageModifier = 100.0f / (100.0f + ((float)defender.Chest.Item.DamageReduction * (1.0f - ((float)attackWeapon.ArmorPenetrationPercent / 100.0f)) - (float)attackWeapon.ArmorPenetrationFlat));
             defender.Stats.Health -= (int)(damage * damageModifier);
 
             //Console.WriteLine($"Damage: {damage}, Damage modifier: {damageModifier}");
